requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
PyJWT>=2.7.0
openai==1.55.3

numpy==2.1.2
faiss-cpu==1.13.1